// src/AppState/ProfileState/Enums/TargetProperies.ts
/** 
 * @TargetProperties
 * Enum provides list of target prop names for change it in reducer
*/
export enum TargetProperties {
    authenticationMethod = 'authenticationMethod',
    creationDate = 'creationDate',
    creator = 'creator',
    fullName = 'fullName',
    id = 'id',
    isActive = 'isActive',
    IsDisabled = 'IsDisabled',
    isSystemAdministrator = 'isSystemAdministrator',
    language = 'language',
    lastLoginDate = 'lastLoginDate',
    lastWriteDate = 'lastWriteDate',
    mbox = 'mbox',
    origin = 'origin',
    role = 'role',
    username = 'username',
    department = 'department',
    manager = 'manager',
    office = 'office',
    phone = 'phone',
    skype = 'skype',
    title = 'title'
};